move LSP Payload to separate mod
